requirements.txt
flask>=3.0.0
python-dotenv>=1.0.0
openai>=1.0.0
scikit-learn>=1.3.0
pandas>=2.0.0
joblib>=1.3.0
sqlalchemy>=2.0.0
